+++ lib/monty_backend_spi.dart
@@ -4,4 +4,4 @@
 /// All types are now in dart_monty_core.
 library;
 
-export 'package:dart_monty_core/dart_monty_core.dart' hide OsCallException;
+export 'package:dart_monty_core/dart_monty_core.dart';